Repot a bit of democracy code
